Spike: runloop refactor
events_loop::runloop is now responsible for the entire Cocoa runloop cycle.
It's a simple blocking implementation (which fails during a resize), but
it's an API that maps cleanly to a coroutine-based implementation.

NSEvent concerns are now separated into events_loop::nsevent.
