+++ src/platform/macos/mod.rs
@@ -42,11 +42,3 @@ mod monitor;
 mod window;
 
 mod timer;
-
-#[cfg(not(feature="context"))]
-mod send_event;
-
-#[cfg(feature="context")]
-mod send_event_context;
-#[cfg(feature="context")]
-use self::send_event_context as send_event;
